+++ crates/translator_new_rml/src/rml_model/v2/io/ref_form_conversion/mod.rs
@@ -38,6 +38,12 @@ impl TryFrom<&ReferenceFormulation>
                                     .to_rcterm()
                             || value
                                 == vocab::rml_io::class::SQL_TABLE
+                                    .to_rcterm()
+                            || value
+                                == vocab::rml_io::class::SQL2008_QUERY
+                                    .to_rcterm()
+                            || value
+                                == vocab::rml_io::class::SQL2008_TABLE
                                     .to_rcterm() =>
                     {
                         Ok(formats::ReferenceFormulation::SQLQuery)